+++ rio_tiler/reader.py
@@ -15,12 +15,12 @@ from rasterio.vrt import WarpedVRT
 from rasterio.warp import transform as transform_coords
 from rasterio.warp import transform_bounds
 
-from .constants import WGS84_CRS
-from .errors import InvalidBufferSize, PointOutsideBounds, TileOutsideBounds
-from .models import ImageData, PointData
-from .types import BBox, DataMaskType, Indexes, NoData
-from .utils import _requested_tile_aligned_with_internal_tile as is_aligned
-from .utils import get_vrt_transform, has_alpha_band, non_alpha_indexes
+from rio_tiler.constants import WGS84_CRS
+from rio_tiler.errors import InvalidBufferSize, PointOutsideBounds, TileOutsideBounds
+from rio_tiler.models import ImageData, PointData
+from rio_tiler.types import BBox, DataMaskType, Indexes, NoData
+from rio_tiler.utils import _requested_tile_aligned_with_internal_tile as is_aligned
+from rio_tiler.utils import get_vrt_transform, has_alpha_band, non_alpha_indexes
 
 
 def _get_width_height(max_size, dataset_height, dataset_width) -> Tuple[int, int]: